修改访客获取咨询师

+++ platform/src/main/java/com/eewoo/platform/pojo/vo/response/CounselorResponse.java
@@ -25,7 +25,7 @@ public class CounselorResponse {
     String email;
     String workPlace;
     String title;
-    int sessionCount;
-    int sessionScore;
+    Integer sessionCount;
+    Double sessionScore;
 
 }
